feat(feature): update profile picture feature - 루틴 프로필 사진 등록하기 화면을 수정합니다.
