+++ feature/update-routine/profile-picture/src/main/kotlin/com/gradation/lift/feature/updateRoutine/profilePicture/ui/ProfilePictureScreen.kt
@@ -1,14 +1,19 @@
 package com.gradation.lift.feature.updateRoutine.profilePicture.ui
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import com.gradation.lift.designsystem.component.text.LiftMultiStyleText
+import com.gradation.lift.designsystem.component.text.LiftTextStyle
+import com.gradation.lift.designsystem.component.text.TextWithStyle
 import com.gradation.lift.designsystem.component.topBar.LiftTopBar
+import com.gradation.lift.designsystem.theme.LiftTheme
 import com.gradation.lift.feature.updateRoutine.common.data.state.CurrentRoutineSetRoutineState
-import com.gradation.lift.feature.updateRoutine.profilePicture.ui.component.LoadingRoutineProfilePictureList
 import com.gradation.lift.feature.updateRoutine.profilePicture.data.state.RoutineSetPictureUiState
-import com.gradation.lift.feature.updateRoutine.profilePicture.ui.component.RoutineProfilePictureList
 import com.gradation.lift.model.model.routine.RoutineSetRoutine
 
 
@@ -23,20 +28,38 @@ fun ProfilePictureScreen(
     Scaffold(
         topBar = {
             LiftTopBar(
-                title = "프로필 등록하기",
+                title = "프로필 사진 등록하기",
                 onClick = navigateProfilePictureToRoutineSetInUpdateRoutineGraph,
             )
         }
     ) { padding ->
         Column(
-            modifier = modifier.padding(padding)
+            modifier = modifier
+                .fillMaxSize()
+                .background(LiftTheme.colorScheme.no5)
+                .padding(padding)
+                .padding(
+                    start = LiftTheme.space.space20,
+                    end = LiftTheme.space.space20,
+                    top = LiftTheme.space.space36
+                ),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.spacedBy(LiftTheme.space.space36)
         ) {
+            LiftMultiStyleText(
+                modifier = modifier,
+                defaultColor = LiftTheme.colorScheme.no11,
+                defaultTextStyle = LiftTextStyle.No1,
+                textAlign = TextAlign.Center,
+                textWithStyleList = listOf(
+                    TextWithStyle(text = "프로필 사진을", color = LiftTheme.colorScheme.no4),
+                    TextWithStyle(text = " 선택해주세요"),
+                )
+            )
+
             when (routineSetPictureUiState) {
                 RoutineSetPictureUiState.Fail -> {}
-                RoutineSetPictureUiState.Loading -> {
-                    LoadingRoutineProfilePictureList(modifier)
-                }
-
+                RoutineSetPictureUiState.Loading -> {}
                 is RoutineSetPictureUiState.Success -> {
                     RoutineProfilePictureList(
                         modifier,
